requirements.txt
asyncio
websockets>=11.0.3
aiohttp>=3.8.5
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
opencv-python>=4.8.1.78
numpy>=1.24.3
requests>=2.31.0